// src/domain/recommendation/models/model.ts
export interface SubwayStation {
  name: string;
  description: string;
  tags: string[];
}

export type RECOMMEND_SUBWAY_STATIONS =  SubwayStation[];

export type PlaceData = {
  address_name: string;
  category_group_code: string;
  category_group_name: string;
  category_name: string;
  distance: string;
  id: string;
  phone: string;
  place_name: string;
  place_url: string;
  road_address_name: string;
  x: string;
  y: string;
};